+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2] - (in progress)
+
+### Added
+- **Collections Creation**:
+  - Added option to create collections from synced catalogs
+
+### Changed
+- Disabled Mirror Sync option if catalog is not marked for syncing to prevent confusion
+
 ## [1.2.1] - 2026-04-14
 
 ### Added